+++ src/main/java/com/ureka/team3/utong_scheduler/product/coupon/service/UserCouponServiceImpl.java
@@ -1,7 +1,7 @@
 package com.ureka.team3.utong_scheduler.product.coupon.service;
 
 import com.ureka.team3.utong_scheduler.product.coupon.repository.UserCouponRepository;
-import com.ureka.team3.utong_scheduler.product.dto.ExpirationResult;
+import com.ureka.team3.utong_scheduler.product.global.dto.ExpirationResult;
 import com.ureka.team3.utong_scheduler.product.global.config.ProductBatchPolicy;
 import com.ureka.team3.utong_scheduler.product.global.enums.ProductType;
 import lombok.RequiredArgsConstructor;
