+++ src/main/java/com/ureka/team3/utong_scheduler/product/gift/service/UserGifticonServiceImpl.java
@@ -1,6 +1,6 @@
 package com.ureka.team3.utong_scheduler.product.gift.service;
 
-import com.ureka.team3.utong_scheduler.product.dto.ExpirationResult;
+import com.ureka.team3.utong_scheduler.product.global.dto.ExpirationResult;
 import com.ureka.team3.utong_scheduler.product.gift.repository.UserGifticonRepository;
 import com.ureka.team3.utong_scheduler.product.global.config.ProductBatchPolicy;
 import com.ureka.team3.utong_scheduler.product.global.enums.ProductType;
